refactor (imageUrl & trailsApi): rename utils/base.js to imageUrl, move trailsApi file path
