+++ src/components/front/Index/BulletinBoard.vue
@@ -94,7 +94,7 @@
 import { ref, computed, onMounted } from 'vue'
 import { RouterLink } from 'vue-router'
 import IconTitle from '@/components/front/base/IconTitle.vue'
-import { fetchTrailsNewsData } from '@/data/api/trailsApi'
+import { fetchTrailsNewsData } from '@/api/trailsApi'
 import { newsType } from '@/data/newsType.js'
 
 const allTailsNews = ref([])
@@ -129,3 +129,4 @@ const trailsNews = computed(() => {
   return result
 })
 </script>
+@/api/trailsApi
