+++ src/views/front/TrailConditionReportView.vue
@@ -60,7 +60,7 @@
 import { onMounted, computed, ref } from 'vue'
 import IconTitle from '@/components/front/base/IconTitle.vue'
 import RoadConditionModal from '@/components/front/base/RoadConditionModal.vue'
-import { fetchTrailsNewsData } from '@/data/api/trailsApi'
+import { fetchTrailsNewsData } from '@/api/trailsApi'
 import { newsType } from '@/data/newsType.js'
 
 import { useMediaQuery } from '@vueuse/core'
